requirements.txt
alembic==1.13.1
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
ecdsa==0.18.0
fastapi==0.108.0
fastapi-utils==0.2.1
filelock==3.13.3
fonttools==4.51.0
fsspec==2024.3.1
h11==0.14.0
idna==3.6
Jinja2==3.1.3
jwt==1.3.1
kiwisolver==1.4.5
Mako==1.3.2
MarkupSafe==2.1.5
matplotlib==3.8.4
minio==7.2.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.1
pillow==10.3.0
psutil==5.9.8
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pyasn1==0.6.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==1.10.15
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.52
starlette==0.32.0.post1
sympy==1.12
thop==0.1.1.post2209072238
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.1.43
urllib3==2.2.1
uvicorn==0.29.0
